requirements.txt
Django~=5.0.2
numpy~=1.26.4
pandas~=2.2.0
openpyxl~=3.1.2
python-dotenv~=1.0.1
scikit-learn~=1.4.0